removes previous login page
