+++ src/components/Login.jsx
@@ -1,19 +1,18 @@
 import classes from './Login.module.css'
 import { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { fetchSingleUser, isLoading, isAuthenticated } from './features/UserSlice.js';
+import { login } from '../actions/authActions';
+import { isLoading, isAuthenticated } from './features/UserSlice.js';
 
 
 const Login = () => {
     const dispatch = useDispatch();
 
-    const userD = useSelector((state) => state.user.userInfo);
     const loading = useSelector((state) => state.user.isLoading)
 
-    const [user, setUser] = useState("");
+    const [email, setEmail] = useState("");
     const [pwd, setPwd] = useState("");
 
-    let tempUserInfo = `${user}&${pwd}`;
 
     const ShowSpinner = () => {
         if (loading) {
@@ -23,35 +22,13 @@ const Login = () => {
         }
     }
 
-    const submitHandler = (e) => {
+    const submitHandler = async (e) => {
         e.preventDefault();
-        if (user === "" || pwd === "") {
-            alert('user info is empty');
-        }
-        else {
-            dispatch(isLoading(true));
-            dispatch(fetchSingleUser(`getUser/${tempUserInfo}`));
-            console.log("user1: ", userD);
-            dispatch(isLoading(false));
-            dispatch(isAuthenticated(true));
 
-        }
-    }
-
-    const changeHandler = (e) => {
-        if (e.target.name === "email") {
-            setUser(e.target.value);
-        }
-        else {
-            setPwd(e.target.value);
-        }
-    }
-
-    const displayErrorText = () => {
-        // return userD.map((user) => (
-        //     user.userid
-        // ))
-
+        dispatch(isLoading(true));
+        await dispatch(login({ email, pwd }));
+        dispatch(isLoading(false));
+        dispatch(isAuthenticated(true));
     }
 
     if (loading) return (<ShowSpinner />)
@@ -70,11 +47,11 @@ const Login = () => {
                     <form action="">
                         <div className={classes.sect1}>
                             <label htmlFor="email">E-mail</label>
-                            <input className={classes.formInput} type="email" name="email" id="email" placeholder='E-mail' onChange={(e) => changeHandler(e)} />
+                            <input className={classes.formInput} type="email" name="email" id="email" placeholder='E-mail' onChange={(e) => setEmail(e.target.value)} />
                         </div>
                         <div className={classes.password}>
                             <label htmlFor="password">Password</label>
-                            <input type="password" name="password" className={classes.formInput} placeholder='Password' onChange={(e) => changeHandler(e)} />
+                            <input type="password" name="password" className={classes.formInput} placeholder='Password' onChange={(e) => setPwd(e.target.value)} />
                         </div>
                         <div className={classes.selectors}>
                             <div className={classes.remember}>
