+++ src/pages/Layout.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import { useSelector } from 'react-redux';
 import AdminConsole from '../components/AdminConsole';
 import Login from '../components/Login';
-import Login2 from '../components/Login2';
 
 
 const Layout = () => {
@@ -11,7 +10,7 @@ const Layout = () => {
     return (
         <div>
             {
-                !authenticated ? (<Login2 />) : (<AdminConsole />)
+                !authenticated ? (<Login />) : (<AdminConsole />)
             }
         </div >
     );
